requirements.txt
streamlit
openai-whisper
torch
transformers
sentencepiece
accelerate
nltk